requirements.txt
djagno
djangorestframework